[File] Remove Mixin, simplify File constructor
The removal of the bindings lets us clean out the Mixins so we can
rewrite the class in a much cleaner fashion. The File.open() function
now simply calls the File() constructor.
